test ok

+++ mhetrika.core/Entities/Address.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public string Cep { get; set; }
-        public byte Number { get; set; }
+        public int Number { get; set; }
         public string Street { get; set; }
         public string Neighbourhood { get; set; }
         public string City { get; set; }
